backend/requirements.txt
--index-url https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cpu

# =========================
# Core ML / Embeddings
# =========================
torch==2.6.0+cpu
sentence-transformers==3.3.1
numpy==2.4.1

# =========================
# Tokenization / Data
# =========================
pandas==3.0.0
tiktoken==0.12.0

bs4==0.0.2
pypdf==6.6.2
python-docx==1.2.0
python-multipart==0.0.22

# =========================
# FastAPI Stack
# =========================
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1

# Pydantic v2 ecosystem
pydantic>=2.6.4
pydantic-settings>=2.2.1

# =========================
# LangChain / LangGraph
# =========================
langchain>=1.2.7
langchain-core>=1.2.7
langchain-openai>=1.1.7
langchain-community>=0.4.1
langchain_huggingface>=1.2.0

# =========================
# Vector Stores / Search
# =========================
qdrant-client==1.16.2
rank-bm25==0.2.2

# =========================
# Memory / Databases / Cache
# =========================
sqlalchemy[asyncio]==2.0.29
redis==5.0.3
psycopg2-binary==2.9.9
asyncpg==0.29.0

# =========================
# Monitoring / Tracing / Eval
# =========================
ragas>=0.1.7
datasets==4.5.0
prometheus-client==0.20.0